Extract base64 image widget logic to a shared module
